+++ test/cli-command-export/test.mjs
@@ -1,12 +1,26 @@
 import { expect } from '@playwright/test';
-import { test, getTestName, execPlaywrightTest, BDDGEN_CMD } from '../helpers.mjs';
+import { test, TestDir, execPlaywrightTest, BDDGEN_CMD } from '../helpers.mjs';
 
-test(getTestName(import.meta), (t) => {
-  const stdout = execPlaywrightTest(t.name, `${BDDGEN_CMD} export`);
-  expect(stdout).toContain('List of all steps found by config: playwright.config.ts');
+const testDir = new TestDir(import.meta);
+
+test(`${testDir.name} (all steps)`, () => {
+  const stdout = execPlaywrightTest(testDir.name, `${BDDGEN_CMD} export`);
+  expect(stdout).toContain('Using config: playwright.config.ts');
+  expect(stdout).toContain('List of all steps (5):');
+
   expect(stdout).toContain('* Given I am on todo page');
   expect(stdout).toContain('* When I add todo {string}');
   expect(stdout).toContain('* Then visible todos count is (\\d+)');
+  expect(stdout).toContain('* When some step');
+  expect(stdout).toContain('* Given I am on another todo page');
+});
+
+test(`${testDir.name} (unused steps)`, () => {
+  const stdout = execPlaywrightTest(testDir.name, `${BDDGEN_CMD} export --unused-steps`);
+  expect(stdout).toContain('Using config: playwright.config.ts');
+  expect(stdout).toContain(`List of unused steps (3):`);
+
+  expect(stdout).toContain('* When I add todo {string}');
   expect(stdout).toContain('* When some step');
   expect(stdout).toContain('* Given I am on another todo page');
 });